Fixed increment marks, slight changes in object values and removed duplicate properties

+++ CA_Charts/classes/BarChart.js
@@ -98,7 +98,7 @@ class BarChart{
                 textSize(this.fontSize);
                 stroke(this.fontSize/3);
                 fill(this.textCol);
-                text((max(this.cleaner.map(row => row[this.yVal]))*(i/this.incrementNum)),-this.incrementWidth*2,-(i*(this.chartHeight/this.incrementNum)))
+                text(int(max(this.cleaner.map(row => row[this.yVal]))*(i/this.incrementNum)),-this.incrementWidth*2,-(i*(this.chartHeight/this.incrementNum)))
             }
         pop();
     }
